Move files from triple-web to tds-widget
