+++ packages/tds-widget/src/index.ts
@@ -22,6 +22,7 @@ export * from './recommended-contents'
 export * from './replies'
 export * from './resource-list-elements'
 export * from './review'
+export * from './scrap'
 export * from './scrap-button'
 export * from './search'
 export * from './social-reviews'
