+++ packages/tds-widget/src/scrap/constants.ts
