+++ packages/tds-widget/src/scrap/types.ts
