+++ packages/triple-web/src/hooks/index.ts
@@ -5,4 +5,3 @@ export * from './modal'
 export * from './session'
 export * from './user-agent'
 export * from './hash-router'
-export * from './scrap'
