+++ packages/triple-web/src/types/index.ts
@@ -6,4 +6,3 @@ export * from './geolocation'
 export * from './modal'
 export * from './session'
 export * from './user-agent'
-export * from './scrap'
